Move node storage into arch.
