+++ fpga_interchange/arch.h
@@ -35,745 +35,10 @@
 #include "dedicated_interconnect.h"
 #include "site_router.h"
 #include "lookahead.h"
+#include "chipdb.h"
+#include "arch_iterators.h"
 
 NEXTPNR_NAMESPACE_BEGIN
-
-/**** Everything in this section must be kept in sync with chipdb.py ****/
-
-// Flattened site indexing.
-//
-// To enable flat BelId.z spaces, every tile and sites within that tile are
-// flattened.
-//
-// This has implications on BelId's, WireId's and PipId's.
-// The flattened site space works as follows:
-//  - Objects that belong to the tile are first.  BELs are always part of Sites,
-//    so no BEL objects are in this category.
-//  - All site alternative modes are exposed as a "full" site.
-//  - Each site appends it's BEL's, wires (site wires) and PIP's.
-//   - Sites add two types of pips.  Sites will add pip data first for site
-//     pips, and then for site pin edges.
-//     1. The first type is site pips, which connect site wires to other site
-//        wires.
-//     2. The second type is site pin edges, which connect site wires to tile
-//        wires (or vise-versa).
-
-NPNR_PACKED_STRUCT(struct BelInfoPOD {
-    int32_t name;       // bel name (in site) constid
-    int32_t type;       // Type name constid
-    int32_t bel_bucket; // BEL bucket constid.
-
-    int32_t num_bel_wires;
-    RelPtr<int32_t> ports; // port name constid
-    RelPtr<int32_t> types; // port type (IN/OUT/BIDIR)
-    RelPtr<int32_t> wires; // connected wire index in tile, or -1 if NA
-
-    int16_t site;
-    int16_t site_variant; // some sites have alternative types
-    int16_t category;
-    int8_t synthetic;
-    int8_t lut_element;
-
-    RelPtr<int32_t> pin_map; // Index into CellMapPOD::cell_bel_map
-});
-
-enum BELCategory
-{
-    // BEL is a logic element
-    BEL_CATEGORY_LOGIC = 0,
-    // BEL is a site routing mux
-    BEL_CATEGORY_ROUTING = 1,
-    // BEL is a site port, e.g. boundry between site and routing graph.
-    BEL_CATEGORY_SITE_PORT = 2
-};
-
-NPNR_PACKED_STRUCT(struct BelPortPOD {
-    int32_t bel_index;
-    int32_t port;
-});
-
-NPNR_PACKED_STRUCT(struct TileWireInfoPOD {
-    int32_t name; // wire name constid
-
-    // Pip index inside tile
-    RelSlice<int32_t> pips_uphill;
-
-    // Pip index inside tile
-    RelSlice<int32_t> pips_downhill;
-
-    // Bel index inside tile
-    RelSlice<BelPortPOD> bel_pins;
-
-    int16_t site;         // site index in tile
-    int16_t site_variant; // site variant index in tile
-});
-
-NPNR_PACKED_STRUCT(struct PipInfoPOD {
-    int32_t src_index, dst_index;
-    int16_t site;         // site index in tile
-    int16_t site_variant; // site variant index in tile
-    int16_t bel;          // BEL this pip belongs to if site pip.
-    int16_t extra_data;
-    RelSlice<int32_t> pseudo_cell_wires;
-});
-
-NPNR_PACKED_STRUCT(struct ConstraintTagPOD {
-    int32_t tag_prefix;       // constid
-    int32_t default_state;    // constid
-    RelSlice<int32_t> states; // constid
-});
-
-NPNR_PACKED_STRUCT(struct LutBelPOD {
-    uint32_t name;          // constid
-    RelSlice<int32_t> pins; // constid
-    uint32_t low_bit;
-    uint32_t high_bit;
-});
-
-NPNR_PACKED_STRUCT(struct LutElementPOD {
-    int32_t width;
-    RelSlice<LutBelPOD> lut_bels;
-});
-
-NPNR_PACKED_STRUCT(struct TileTypeInfoPOD {
-    int32_t name; // Tile type constid
-
-    RelSlice<BelInfoPOD> bel_data;
-
-    RelSlice<TileWireInfoPOD> wire_data;
-
-    RelSlice<PipInfoPOD> pip_data;
-
-    RelSlice<ConstraintTagPOD> tags;
-
-    RelSlice<LutElementPOD> lut_elements;
-
-    RelSlice<int32_t> site_types; // constid
-});
-
-NPNR_PACKED_STRUCT(struct SiteInstInfoPOD {
-    RelPtr<char> name;
-    RelPtr<char> site_name;
-
-    // Which site type is this site instance?
-    // constid
-    int32_t site_type;
-});
-
-NPNR_PACKED_STRUCT(struct TileInstInfoPOD {
-    // Name of this tile.
-    RelPtr<char> name;
-
-    // Index into root.tile_types.
-    int32_t type;
-
-    // This array is root.tile_types[type].site_types.size() long.
-    // Index into root.sites
-    RelSlice<int32_t> sites;
-
-    // Number of tile wires; excluding any site-internal wires
-    // which come after general wires and are not stored here
-    // as they will never be nodal
-    // -1 if a tile-local wire; node index if nodal wire
-    RelSlice<int32_t> tile_wire_to_node;
-});
-
-NPNR_PACKED_STRUCT(struct TileWireRefPOD {
-    int32_t tile;
-    int32_t index;
-});
-
-NPNR_PACKED_STRUCT(struct NodeInfoPOD { RelSlice<TileWireRefPOD> tile_wires; });
-
-NPNR_PACKED_STRUCT(struct CellBelPinPOD {
-    int32_t cell_pin; // constid
-    int32_t bel_pin;  // constid
-});
-
-NPNR_PACKED_STRUCT(struct ParameterPinsPOD {
-    int32_t key;   // constid
-    int32_t value; // constid
-    RelSlice<CellBelPinPOD> pins;
-});
-
-NPNR_PACKED_STRUCT(struct CellConstraintPOD {
-    int32_t tag;              // Tag index
-    int32_t constraint_type;  // Constraint::ConstraintType
-    RelSlice<int32_t> states; // State indicies
-});
-
-NPNR_PACKED_STRUCT(struct CellBelMapPOD {
-    RelSlice<CellBelPinPOD> common_pins;
-    RelSlice<ParameterPinsPOD> parameter_pins;
-    RelSlice<CellConstraintPOD> constraints;
-});
-
-NPNR_PACKED_STRUCT(struct LutCellPOD {
-    int32_t cell;                 // constid
-    RelSlice<int32_t> input_pins; // constids
-    int32_t parameter;
-});
-
-NPNR_PACKED_STRUCT(struct CellMapPOD {
-    // Cell names supported in this arch.
-    RelSlice<int32_t> cell_names;       // constids
-    RelSlice<int32_t> cell_bel_buckets; // constids
-
-    RelSlice<CellBelMapPOD> cell_bel_map;
-
-    RelSlice<LutCellPOD> lut_cells;
-});
-
-NPNR_PACKED_STRUCT(struct PackagePinPOD {
-    int32_t package_pin; // constid
-    int32_t site;        // constid
-    int32_t bel;         // constid
-});
-
-NPNR_PACKED_STRUCT(struct PackagePOD {
-    int32_t package; // constid
-    RelSlice<PackagePinPOD> pins;
-});
-
-NPNR_PACKED_STRUCT(struct ConstantsPOD {
-    // Cell type and port for the GND and VCC global source.
-    int32_t gnd_cell_name; // constid
-    int32_t gnd_cell_port; // constid
-
-    int32_t vcc_cell_name; // constid
-    int32_t vcc_cell_port; // constid
-
-    int32_t gnd_bel_tile;
-    int32_t gnd_bel_index;
-    int32_t gnd_bel_pin; // constid
-
-    int32_t vcc_bel_tile;
-    int32_t vcc_bel_index;
-    int32_t vcc_bel_pin; // constid
-
-    // Name to use for the global GND constant net
-    int32_t gnd_net_name; // constid
-
-    // Name to use for the global VCC constant net
-    int32_t vcc_net_name; // constid
-});
-
-NPNR_PACKED_STRUCT(struct ChipInfoPOD {
-    RelPtr<char> name;
-    RelPtr<char> generator;
-
-    int32_t version;
-    int32_t width, height;
-
-    RelSlice<TileTypeInfoPOD> tile_types;
-    RelSlice<SiteInstInfoPOD> sites;
-    RelSlice<TileInstInfoPOD> tiles;
-    RelSlice<NodeInfoPOD> nodes;
-    RelSlice<PackagePOD> packages;
-
-    // BEL bucket constids.
-    RelSlice<int32_t> bel_buckets;
-
-    RelPtr<CellMapPOD> cell_map;
-    RelPtr<ConstantsPOD> constants;
-
-    // Constid string data.
-    RelPtr<RelSlice<RelPtr<char>>> constids;
-});
-
-/************************ End of chipdb section. ************************/
-
-inline const TileTypeInfoPOD &tile_info(const ChipInfoPOD *chip_info, int32_t tile)
-{
-    return chip_info->tile_types[chip_info->tiles[tile].type];
-}
-
-template <typename Id> const TileTypeInfoPOD &loc_info(const ChipInfoPOD *chip_info, Id &id)
-{
-    return chip_info->tile_types[chip_info->tiles[id.tile].type];
-}
-
-inline const BelInfoPOD &bel_info(const ChipInfoPOD *chip_info, BelId bel)
-{
-    NPNR_ASSERT(bel != BelId());
-    return loc_info(chip_info, bel).bel_data[bel.index];
-}
-
-inline const PipInfoPOD &pip_info(const ChipInfoPOD *chip_info, PipId pip)
-{
-    NPNR_ASSERT(pip != PipId());
-    return loc_info(chip_info, pip).pip_data[pip.index];
-}
-
-inline const SiteInstInfoPOD &site_inst_info(const ChipInfoPOD *chip_info, int32_t tile, int32_t site)
-{
-    return chip_info->sites[chip_info->tiles[tile].sites[site]];
-}
-
-struct BelIterator
-{
-    const ChipInfoPOD *chip;
-    int cursor_index;
-    int cursor_tile;
-
-    BelIterator operator++()
-    {
-        cursor_index++;
-        while (cursor_tile < chip->tiles.ssize() && cursor_index >= tile_info(chip, cursor_tile).bel_data.ssize()) {
-            cursor_index = 0;
-            cursor_tile++;
-        }
-        return *this;
-    }
-    BelIterator operator++(int)
-    {
-        BelIterator prior(*this);
-        ++(*this);
-        return prior;
-    }
-
-    bool operator!=(const BelIterator &other) const
-    {
-        return cursor_index != other.cursor_index || cursor_tile != other.cursor_tile;
-    }
-
-    bool operator==(const BelIterator &other) const
-    {
-        return cursor_index == other.cursor_index && cursor_tile == other.cursor_tile;
-    }
-
-    BelId operator*() const
-    {
-        BelId ret;
-        ret.tile = cursor_tile;
-        ret.index = cursor_index;
-        return ret;
-    }
-};
-
-struct BelRange
-{
-    BelIterator b, e;
-    BelIterator begin() const { return b; }
-    BelIterator end() const { return e; }
-};
-
-struct FilteredBelIterator
-{
-    std::function<bool(BelId)> filter;
-    BelIterator b, e;
-
-    FilteredBelIterator operator++()
-    {
-        ++b;
-        while (b != e) {
-            if (filter(*b)) {
-                break;
-            }
-
-            ++b;
-        }
-        return *this;
-    }
-
-    bool operator!=(const FilteredBelIterator &other) const
-    {
-        NPNR_ASSERT(e == other.e);
-        return b != other.b;
-    }
-
-    bool operator==(const FilteredBelIterator &other) const
-    {
-        NPNR_ASSERT(e == other.e);
-        return b == other.b;
-    }
-
-    BelId operator*() const
-    {
-        BelId bel = *b;
-        NPNR_ASSERT(filter(bel));
-        return bel;
-    }
-};
-
-struct FilteredBelRange
-{
-    FilteredBelRange(BelIterator bel_b, BelIterator bel_e, std::function<bool(BelId)> filter)
-    {
-        b.filter = filter;
-        b.b = bel_b;
-        b.e = bel_e;
-
-        if (b.b != b.e && !filter(*b.b)) {
-            ++b;
-        }
-
-        e.b = bel_e;
-        e.e = bel_e;
-
-        if (b != e) {
-            NPNR_ASSERT(filter(*b.b));
-        }
-    }
-
-    FilteredBelIterator b, e;
-    FilteredBelIterator begin() const { return b; }
-    FilteredBelIterator end() const { return e; }
-};
-
-// -----------------------------------------------------------------------
-
-// Iterate over TileWires for a wire (will be more than one if nodal)
-struct TileWireIterator
-{
-    const ChipInfoPOD *chip;
-    WireId baseWire;
-    int cursor = -1;
-
-    void operator++() { cursor++; }
-
-    bool operator==(const TileWireIterator &other) const { return cursor == other.cursor; }
-    bool operator!=(const TileWireIterator &other) const { return cursor != other.cursor; }
-
-    // Returns a *denormalised* identifier always pointing to a tile wire rather than a node
-    WireId operator*() const
-    {
-        if (baseWire.tile == -1) {
-            WireId tw;
-            const auto &node_wire = chip->nodes[baseWire.index].tile_wires[cursor];
-            tw.tile = node_wire.tile;
-            tw.index = node_wire.index;
-            return tw;
-        } else {
-            return baseWire;
-        }
-    }
-};
-
-struct TileWireRange
-{
-    TileWireIterator b, e;
-    TileWireIterator begin() const { return b; }
-    TileWireIterator end() const { return e; }
-};
-
-inline WireId canonical_wire(const ChipInfoPOD *chip_info, int32_t tile, int32_t wire)
-{
-    WireId id;
-
-    if (wire >= chip_info->tiles[tile].tile_wire_to_node.ssize()) {
-        // Cannot be a nodal wire
-        id.tile = tile;
-        id.index = wire;
-    } else {
-        int32_t node = chip_info->tiles[tile].tile_wire_to_node[wire];
-        if (node == -1) {
-            // Not a nodal wire
-            id.tile = tile;
-            id.index = wire;
-        } else {
-            // Is a nodal wire, set tile to -1
-            id.tile = -1;
-            id.index = node;
-        }
-    }
-
-    return id;
-}
-
-// -----------------------------------------------------------------------
-
-struct WireIterator
-{
-    const ChipInfoPOD *chip;
-    int cursor_index = 0;
-    int cursor_tile = -1;
-
-    WireIterator operator++()
-    {
-        // Iterate over nodes first, then tile wires that aren't nodes
-        do {
-            cursor_index++;
-            if (cursor_tile == -1 && cursor_index >= chip->nodes.ssize()) {
-                cursor_tile = 0;
-                cursor_index = 0;
-            }
-            while (cursor_tile != -1 && cursor_tile < chip->tiles.ssize() &&
-                   cursor_index >= chip->tile_types[chip->tiles[cursor_tile].type].wire_data.ssize()) {
-                cursor_index = 0;
-                cursor_tile++;
-            }
-
-        } while ((cursor_tile != -1 && cursor_tile < chip->tiles.ssize() &&
-                  cursor_index < chip->tiles[cursor_tile].tile_wire_to_node.ssize() &&
-                  chip->tiles[cursor_tile].tile_wire_to_node[cursor_index] != -1));
-
-        return *this;
-    }
-    WireIterator operator++(int)
-    {
-        WireIterator prior(*this);
-        ++(*this);
-        return prior;
-    }
-
-    bool operator!=(const WireIterator &other) const
-    {
-        return cursor_index != other.cursor_index || cursor_tile != other.cursor_tile;
-    }
-
-    bool operator==(const WireIterator &other) const
-    {
-        return cursor_index == other.cursor_index && cursor_tile == other.cursor_tile;
-    }
-
-    WireId operator*() const
-    {
-        WireId ret;
-        ret.tile = cursor_tile;
-        ret.index = cursor_index;
-        return ret;
-    }
-};
-
-struct WireRange
-{
-    WireIterator b, e;
-    WireIterator begin() const { return b; }
-    WireIterator end() const { return e; }
-};
-
-// -----------------------------------------------------------------------
-struct AllPipIterator
-{
-    const ChipInfoPOD *chip;
-    int cursor_index;
-    int cursor_tile;
-
-    AllPipIterator operator++()
-    {
-        cursor_index++;
-        while (cursor_tile < chip->tiles.ssize() &&
-               cursor_index >= chip->tile_types[chip->tiles[cursor_tile].type].pip_data.ssize()) {
-            cursor_index = 0;
-            cursor_tile++;
-        }
-        return *this;
-    }
-    AllPipIterator operator++(int)
-    {
-        AllPipIterator prior(*this);
-        ++(*this);
-        return prior;
-    }
-
-    bool operator!=(const AllPipIterator &other) const
-    {
-        return cursor_index != other.cursor_index || cursor_tile != other.cursor_tile;
-    }
-
-    bool operator==(const AllPipIterator &other) const
-    {
-        return cursor_index == other.cursor_index && cursor_tile == other.cursor_tile;
-    }
-
-    PipId operator*() const
-    {
-        PipId ret;
-        ret.tile = cursor_tile;
-        ret.index = cursor_index;
-        return ret;
-    }
-};
-
-struct AllPipRange
-{
-    AllPipIterator b, e;
-    AllPipIterator begin() const { return b; }
-    AllPipIterator end() const { return e; }
-};
-
-// -----------------------------------------------------------------------
-
-struct UphillPipIterator
-{
-    const ChipInfoPOD *chip;
-    TileWireIterator twi, twi_end;
-    int cursor = -1;
-
-    void operator++()
-    {
-        cursor++;
-        while (true) {
-            if (!(twi != twi_end))
-                break;
-            WireId w = *twi;
-            auto &tile = chip->tile_types[chip->tiles[w.tile].type];
-            if (cursor < tile.wire_data[w.index].pips_uphill.ssize())
-                break;
-            ++twi;
-            cursor = 0;
-        }
-    }
-    bool operator!=(const UphillPipIterator &other) const { return twi != other.twi || cursor != other.cursor; }
-
-    PipId operator*() const
-    {
-        PipId ret;
-        WireId w = *twi;
-        ret.tile = w.tile;
-        ret.index = chip->tile_types[chip->tiles[w.tile].type].wire_data[w.index].pips_uphill[cursor];
-        return ret;
-    }
-};
-
-struct UphillPipRange
-{
-    UphillPipIterator b, e;
-    UphillPipIterator begin() const { return b; }
-    UphillPipIterator end() const { return e; }
-};
-
-struct DownhillPipIterator
-{
-    const ChipInfoPOD *chip;
-    TileWireIterator twi, twi_end;
-    int cursor = -1;
-
-    int32_t tile;
-    int32_t tile_type;
-    const RelSlice<int32_t> * pips_downhill = nullptr;
-
-    void operator++()
-    {
-        cursor++;
-        while (true) {
-            if (!(twi != twi_end))
-                break;
-
-            if(pips_downhill == nullptr) {
-                WireId w = *twi;
-                tile_type = chip->tiles[w.tile].type;
-                const TileTypeInfoPOD &type = chip->tile_types[tile_type];
-
-                tile = w.tile;
-                pips_downhill = &type.wire_data[w.index].pips_downhill;
-            }
-
-            if (cursor < pips_downhill->ssize())
-                break;
-
-            ++twi;
-            cursor = 0;
-            pips_downhill = nullptr;
-        }
-    }
-    bool operator!=(const DownhillPipIterator &other) const { return twi != other.twi || cursor != other.cursor; }
-
-    PipId operator*() const
-    {
-        PipId ret;
-        ret.tile = tile;
-        ret.index = (*pips_downhill)[cursor];
-        return ret;
-    }
-};
-
-struct DownhillPipRange
-{
-    DownhillPipIterator b, e;
-    DownhillPipIterator begin() const { return b; }
-    DownhillPipIterator end() const { return e; }
-};
-
-struct BelPinIterator
-{
-    const ChipInfoPOD *chip;
-    TileWireIterator twi, twi_end;
-    int cursor = -1;
-
-    void operator++()
-    {
-        cursor++;
-
-        while (twi != twi_end) {
-            WireId w = *twi;
-            auto &tile = tile_info(chip, w.tile);
-            if (cursor < tile.wire_data[w.index].bel_pins.ssize())
-                break;
-
-            ++twi;
-            cursor = 0;
-        }
-    }
-    bool operator!=(const BelPinIterator &other) const { return twi != other.twi || cursor != other.cursor; }
-
-    BelPin operator*() const
-    {
-        BelPin ret;
-        WireId w = *twi;
-        ret.bel.tile = w.tile;
-        ret.bel.index = tile_info(chip, w.tile).wire_data[w.index].bel_pins[cursor].bel_index;
-        ret.pin.index = tile_info(chip, w.tile).wire_data[w.index].bel_pins[cursor].port;
-        return ret;
-    }
-};
-
-struct BelPinRange
-{
-    BelPinIterator b, e;
-    BelPinIterator begin() const { return b; }
-    BelPinIterator end() const { return e; }
-};
-
-struct IdStringIterator : std::iterator<std::forward_iterator_tag,
-                                        /*T=*/IdString,
-                                        /*Distance=*/ptrdiff_t,
-                                        /*pointer=*/IdString *,
-                                        /*reference=*/IdString>
-{
-    const int32_t *cursor;
-
-    void operator++() { cursor += 1; }
-
-    bool operator!=(const IdStringIterator &other) const { return cursor != other.cursor; }
-
-    bool operator==(const IdStringIterator &other) const { return cursor == other.cursor; }
-
-    IdString operator*() const { return IdString(*cursor); }
-};
-
-struct IdStringRange
-{
-    IdStringIterator b, e;
-    IdStringIterator begin() const { return b; }
-    IdStringIterator end() const { return e; }
-};
-
-struct BelBucketIterator
-{
-    IdStringIterator cursor;
-
-    void operator++() { ++cursor; }
-
-    bool operator!=(const BelBucketIterator &other) const { return cursor != other.cursor; }
-
-    bool operator==(const BelBucketIterator &other) const { return cursor == other.cursor; }
-
-    BelBucketId operator*() const
-    {
-        BelBucketId bucket;
-        bucket.name = IdString(*cursor);
-        return bucket;
-    }
-};
-
-struct BelBucketRange
-{
-    BelBucketIterator b, e;
-    BelBucketIterator begin() const { return b; }
-    BelBucketIterator end() const { return e; }
-};
 
 struct ArchArgs
 {
@@ -1762,6 +1027,7 @@ struct Arch : ArchAPI<ArchRanges>
     IdString id_BUFG;
 
     Lookahead lookahead;
+    mutable RouteNodeStorage node_storage;
 };
 
 NEXTPNR_NAMESPACE_END
