+++ fpga_interchange/site_arch.h
@@ -18,13 +18,22 @@
  *
  */
 
-#include "nextpnr.h"
-#include "log.h"
-
 #ifndef SITE_ARCH_H
 #define SITE_ARCH_H
 
+#include <cstdint>
+#include <unordered_set>
+#include <vector>
+
+#include "nextpnr_namespaces.h"
+#include "nextpnr_types.h"
+#include "chipdb.h"
+#include "log.h"
+#include "arch_iterators.h"
+
 NEXTPNR_NAMESPACE_BEGIN
+
+struct Context;
 
 struct SiteInformation {
     const Context *ctx;
@@ -34,34 +43,17 @@ struct SiteInformation {
     const int32_t site;
     const std::unordered_set<CellInfo *> &cells_in_site;
 
-    SiteInformation(const Context *ctx, int32_t tile, int32_t site, const std::unordered_set<CellInfo *> &cells_in_site) : ctx(ctx), tile(tile), tile_type(ctx->chip_info->tiles[tile].type), site(site), cells_in_site(cells_in_site) {}
+    SiteInformation(const Context *ctx, int32_t tile, int32_t site, const std::unordered_set<CellInfo *> &cells_in_site);
 
-    bool is_wire_in_site(WireId wire) const {
-        if(wire.tile != tile) {
-            return false;
-        }
+    const ChipInfoPOD & chip_info() const;
 
-        return ctx->wire_info(wire).site == site;
-    }
+    bool is_wire_in_site(WireId wire) const;
 
-    bool is_bel_in_site(BelId bel) const {
-        if(bel.tile != tile) {
-            return false;
-        }
+    bool is_bel_in_site(BelId bel) const;
 
-        return bel_info(ctx->chip_info, bel).site == site;
+    bool is_pip_part_of_site(PipId pip) const;
 
-    }
-
-    bool is_pip_part_of_site(PipId pip) const {
-        if(pip.tile != tile) {
-            return false;
-        }
-
-        const auto & tile_type_data = ctx->chip_info->tile_types[tile_type];
-        const auto & pip_data = tile_type_data.pip_data[pip.index];
-        return pip_data.site == site;
-    }
+    bool is_site_port(PipId pip) const;
 };
 
 // Site routing needs a modification of the routing graph.  Within the site,
@@ -133,30 +125,54 @@ struct SiteWire {
 
 
     static SiteWire make_site_port(const SiteInformation *site_info, PipId pip, WireId wire) {
-        NPNR_ASSERT(site_info->ctx->is_site_port(pip));
+        const auto & tile_type_data = site_info->chip_info().tile_types[site_info->tile_type];
+        const auto & pip_data = tile_type_data.pip_data[pip.index];
+
+        // This pip should definitely be part of this site
+        NPNR_ASSERT(pip_data.site == site_info->site);
 
         SiteWire out;
         out.wire = wire;
 
-        WireId src_wire = site_info->ctx->getPipSrcWire(pip);
-        WireId dst_wire = site_info->ctx->getPipDstWire(pip);
-        if(wire == src_wire) {
-            if(site_info->is_wire_in_site(src_wire)) {
-                NPNR_ASSERT(!site_info->is_wire_in_site(dst_wire));
+        const auto & src_data = tile_type_data.wire_data[pip_data.src_index];
+        const auto & dst_data = tile_type_data.wire_data[pip_data.dst_index];
+
+        bool wire_is_src;
+        if(wire.tile == site_info->tile) {
+            // `wire` argument is **not** a node, so simple equality works.
+            wire_is_src = wire.index == pip_data.src_index;
+            if(!wire_is_src) {
+                NPNR_ASSERT(wire.index == pip_data.dst_index);
+            }
+        } else {
+            // `wire` argument **is** a node, so the site wire is not the
+            // `wire`.
+            NPNR_ASSERT(wire.tile == -1);
+
+            // If the source wire is not part of the site, then the source
+            // wire is the `wire` argument.
+            wire_is_src = (src_data.site == -1);
+        }
+
+        if(wire_is_src) {
+            if(src_data.site == site_info->site) {
+                NPNR_ASSERT(dst_data.site == -1);
                 out.type = SITE_WIRE;
             } else {
-                NPNR_ASSERT(site_info->is_wire_in_site(dst_wire));
+                NPNR_ASSERT(src_data.site == -1);
+                NPNR_ASSERT(dst_data.site == site_info->site);
                 out.type = SITE_PORT_SOURCE;
                 out.pip = pip;
             }
         } else {
-            if(site_info->is_wire_in_site(dst_wire)) {
-                NPNR_ASSERT(!site_info->is_wire_in_site(src_wire));
-                out.type = SITE_WIRE;
-            } else {
-                NPNR_ASSERT(site_info->is_wire_in_site(src_wire));
+            if(src_data.site == site_info->site) {
+                NPNR_ASSERT(dst_data.site == -1);
                 out.type = SITE_PORT_SINK;
                 out.pip = pip;
+            } else {
+                NPNR_ASSERT(src_data.site == -1);
+                NPNR_ASSERT(dst_data.site == site_info->site);
+                out.type = SITE_WIRE;
             }
         }
 
@@ -195,7 +211,7 @@ struct SitePip {
         SitePip out;
         out.pip = pip;
 
-        if(site_info->ctx->is_site_port(pip)) {
+        if(site_info->is_site_port(pip)) {
             out.type = SITE_PORT;
         } else {
             out.type = SITE_PIP;
@@ -205,8 +221,6 @@ struct SitePip {
 
     static SitePip make(const SiteInformation *site_info, SiteWire src, PipId dst) {
         NPNR_ASSERT(src.type == SiteWire::OUT_OF_SITE_SOURCE);
-        NPNR_ASSERT(site_info->ctx->is_site_port(dst));
-        NPNR_ASSERT(site_info->is_wire_in_site(site_info->ctx->getPipDstWire(dst)));
 
         SitePip out;
         out.type = SOURCE_TO_SITE_PORT;
@@ -217,8 +231,6 @@ struct SitePip {
     }
 
     static SitePip make(const SiteInformation *site_info, PipId src, SiteWire dst) {
-        NPNR_ASSERT(site_info->ctx->is_site_port(src));
-        NPNR_ASSERT(site_info->is_wire_in_site(site_info->ctx->getPipSrcWire(src)));
         NPNR_ASSERT(dst.type == SiteWire::OUT_OF_SITE_SINK);
 
         SitePip out;
@@ -234,12 +246,6 @@ struct SitePip {
         out.type = SITE_PORT_TO_SITE_PORT;
         out.pip = src_pip;
         out.other_pip = dst_pip;
-
-        NPNR_ASSERT(site_info->ctx->is_site_port(src_pip));
-        NPNR_ASSERT(site_info->ctx->is_site_port(dst_pip));
-
-        NPNR_ASSERT(site_info->is_wire_in_site(site_info->ctx->getPipSrcWire(src_pip)));
-        NPNR_ASSERT(site_info->is_wire_in_site(site_info->ctx->getPipDstWire(dst_pip)));
 
         return out;
     }
@@ -326,265 +332,29 @@ struct SiteArch {
     std::vector<SiteWire> out_of_site_sources;
     std::vector<SiteWire> out_of_site_sinks;
 
-    SiteArch(const SiteInformation *site_info) : ctx(site_info->ctx), site_info(site_info) {
-        // Build list of input and output site ports
-        //
-        // FIXME: This doesn't need to be computed over and over, move to
-        // arch/chip db.
-        const TileTypeInfoPOD & tile_type = loc_info(site_info->ctx->chip_info, *site_info);
-        PipId pip;
-        pip.tile = site_info->tile;
-        for(size_t pip_index = 0; pip_index < tile_type.pip_data.size(); ++pip_index) {
-            if(tile_type.pip_data[pip_index].site != site_info->site) {
-                continue;
-            }
+    SiteArch(const SiteInformation *site_info);
 
-            pip.index = pip_index;
-
-            if(!ctx->is_site_port(pip)) {
-                continue;
-            }
-
-            WireId src_wire = ctx->getPipSrcWire(pip);
-            if(site_info->is_wire_in_site(src_wire)) {
-                output_site_ports.push_back(pip);
-            } else {
-                input_site_ports.push_back(pip);
-            }
-        }
-
-        // Create list of out of site sources and sinks.
-
-        for (CellInfo *cell : site_info->cells_in_site) {
-            for (const auto &pin_pair : cell->cell_bel_pins) {
-                const PortInfo &port = cell->ports.at(pin_pair.first);
-                if(port.net != nullptr) {
-                    nets.emplace(port.net, SiteNetInfo{port.net});
-                }
-            }
-        }
-
-        for(auto &net_pair : nets) {
-            NetInfo *net = net_pair.first;
-            SiteNetInfo &net_info = net_pair.second;
-
-            // All nets require drivers
-            NPNR_ASSERT(net->driver.cell != nullptr);
-
-            bool net_driven_out_of_site = false;
-            if(net->driver.cell->bel == BelId()) {
-                // The driver of this site hasn't been placed, so treat it as
-                // out of site.
-                out_of_site_sources.push_back(SiteWire::make(site_info, PORT_OUT, net));
-                net_info.driver = out_of_site_sources.back();
-                net_driven_out_of_site = true;
-            } else {
-                if(!site_info->is_bel_in_site(net->driver.cell->bel)) {
-
-                    // The driver of this site has been placed, it is an out
-                    // of site source.
-                    out_of_site_sources.push_back(
-                            SiteWire::make(site_info, PORT_OUT,
-                                net));
-                    out_of_site_sources.back().wire = ctx->getNetinfoSourceWire(net);
-                    net_info.driver = out_of_site_sources.back();
-
-                    net_driven_out_of_site = true;
-                } else {
-                    net_info.driver = SiteWire::make(site_info, ctx->getNetinfoSourceWire(net));
-                }
-            }
-
-            if(net_driven_out_of_site) {
-                // Because this net is driven from a source out of the site,
-                // no out of site sink is required.
-                continue;
-            }
-
-            // Examine net to determine if it has any users not in this site.
-            bool net_used_out_of_site = false;
-            WireId out_of_site_wire;
-            for(const PortRef & user : net->users) {
-                if(user.cell == nullptr) {
-                    // This is pretty weird!
-                    continue;
-                }
-
-                if(user.cell->bel == BelId()) {
-                    // Because this net has a user that has not been placed,
-                    // and this net is being driven from this site, make sure
-                    // this net can be routed from this site.
-                    net_used_out_of_site = true;
-                    continue;
-                }
-
-                if(!site_info->is_bel_in_site(user.cell->bel)) {
-                    net_used_out_of_site = true;
-                    for(IdString bel_pin : ctx->getBelPinsForCellPin(user.cell, user.port)) {
-                        out_of_site_wire = ctx->getBelPinWire(user.cell->bel, bel_pin);
-                        break;
-                    }
-                }
-            }
-
-            if(net_used_out_of_site) {
-                out_of_site_sinks.push_back(
-                        SiteWire::make(site_info, PORT_IN,
-                            net));
-                if(out_of_site_wire != WireId()) {
-                    out_of_site_sinks.back().wire = out_of_site_wire;
-                }
-
-                net_info.users.emplace(out_of_site_sinks.back());
-            }
-        }
-
-        // At this point all nets have a driver SiteWire, but user SiteWire's
-        // within the site are not present.  Add them now.
-        for(auto &net_pair : nets) {
-            NetInfo *net = net_pair.first;
-            SiteNetInfo &net_info = net_pair.second;
-
-            for(const PortRef & user : net->users) {
-                if(!site_info->is_bel_in_site(user.cell->bel)) {
-                    // Only care about BELs within the site at this point.
-                    continue;
-                }
-
-                for(IdString bel_pin : ctx->getBelPinsForCellPin(user.cell, user.port)) {
-                    SiteWire wire = getBelPinWire(user.cell->bel, bel_pin);
-                    // Don't add users that are trivially routable!
-                    if(wire != net_info.driver) {
-                        if(ctx->debug) {
-                            log_info("Add user %s because it isn't driver %s\n",
-                                    nameOfWire(wire), nameOfWire(net_info.driver));
-                        }
-                        net_info.users.emplace(wire);
-                    }
-                }
-            }
-        }
-
-        for(auto &net_pair : nets) {
-            SiteNetInfo *net_info = &net_pair.second;
-            auto result = wire_to_nets.emplace(net_info->driver, SiteNetMap{net_info, 1});
-            // By this point, trivial congestion at sources should already by
-            // avoided, and there should be no duplicates in the
-            // driver/users data.
-            NPNR_ASSERT(result.second);
-
-            for(const auto &user : net_info->users) {
-                result = wire_to_nets.emplace(user, SiteNetMap{net_info, 1});
-                NPNR_ASSERT(result.second);
-            }
-        }
-    }
-
-    SiteWire getPipSrcWire(const SitePip &site_pip) const {
-        SiteWire site_wire;
-        switch(site_pip.type) {
-        case SitePip::Type::SITE_PIP:
-            return SiteWire::make(site_info, ctx->getPipSrcWire(site_pip.pip));
-        case SitePip::Type::SITE_PORT:
-            return SiteWire::make_site_port(site_info, site_pip.pip, ctx->getPipSrcWire(site_pip.pip));
-        case SitePip::Type::SOURCE_TO_SITE_PORT:
-            NPNR_ASSERT(site_pip.wire.type == SiteWire::OUT_OF_SITE_SOURCE);
-            return site_pip.wire;
-        case SitePip::Type::SITE_PORT_TO_SINK:
-            site_wire = SiteWire::make_site_port(site_info, site_pip.pip, ctx->getPipDstWire(site_pip.pip));
-            NPNR_ASSERT(site_wire.type == SiteWire::SITE_PORT_SINK);
-            return site_wire;
-        case SitePip::Type::SITE_PORT_TO_SITE_PORT:
-            site_wire = SiteWire::make_site_port(site_info, site_pip.pip, ctx->getPipDstWire(site_pip.pip));
-            NPNR_ASSERT(site_wire.type == SiteWire::SITE_PORT_SINK);
-            return site_wire;
-        default:
-            // Unreachable!
-            NPNR_ASSERT(false);
-        }
-    }
-
-    SiteWire getPipDstWire(const SitePip &site_pip) const {
-        SiteWire site_wire;
-        switch(site_pip.type) {
-        case SitePip::Type::SITE_PIP:
-            return SiteWire::make(site_info, ctx->getPipDstWire(site_pip.pip));
-        case SitePip::Type::SITE_PORT:
-            return SiteWire::make_site_port(site_info, site_pip.pip, ctx->getPipDstWire(site_pip.pip));
-        case SitePip::Type::SOURCE_TO_SITE_PORT:
-            site_wire = SiteWire::make_site_port(site_info, site_pip.pip, ctx->getPipSrcWire(site_pip.pip));
-            NPNR_ASSERT(site_wire.type == SiteWire::SITE_PORT_SOURCE);
-            return site_wire;
-        case SitePip::Type::SITE_PORT_TO_SINK:
-            NPNR_ASSERT(site_pip.wire.type == SiteWire::OUT_OF_SITE_SINK);
-            return site_pip.wire;
-        case SitePip::Type::SITE_PORT_TO_SITE_PORT:
-            site_wire = SiteWire::make_site_port(site_info, site_pip.other_pip, ctx->getPipSrcWire(site_pip.other_pip));
-            NPNR_ASSERT(site_wire.type == SiteWire::SITE_PORT_SOURCE);
-            return site_wire;
-        default:
-            // Unreachable!
-            NPNR_ASSERT(false);
-        }
-    }
+    SiteWire getPipSrcWire(const SitePip &site_pip) const;
+    SiteWire getPipDstWire(const SitePip &site_pip) const;
 
     SitePipDownhillRange getPipsDownhill(const SiteWire &site_wire) const;
     SitePipUphillRange getPipsUphill(const SiteWire &site_wire) const;
     SiteWireRange getWires() const;
 
-    SiteWire getBelPinWire(BelId bel, IdString pin) const {
-        WireId wire = ctx->getBelPinWire(bel, pin);
-        return SiteWire::make(site_info, wire);
-    }
+    SiteWire getBelPinWire(BelId bel, IdString pin) const;
+    PortType getBelPinType(BelId bel, IdString pin) const;
 
-    PortType getBelPinType(BelId bel, IdString pin) const {
-        return ctx->getBelPinType(bel, pin);
-    }
+    const char * nameOfWire(const SiteWire &wire) const;
+    const char * nameOfPip(const SitePip &pip) const;
+    const char * nameOfNet(const SiteNetInfo *net) const;
 
-    const char * nameOfWire(const SiteWire &wire) const {
-        switch(wire.type) {
-        case SiteWire::SITE_WIRE:
-            return ctx->nameOfWire(wire.wire);
-        case SiteWire::SITE_PORT_SINK:
-            return ctx->nameOfWire(wire.wire);
-        case SiteWire::SITE_PORT_SOURCE:
-            return ctx->nameOfWire(wire.wire);
-        case SiteWire::OUT_OF_SITE_SOURCE:
-            return "out of site source, implement me!";
-        case SiteWire::OUT_OF_SITE_SINK:
-            return "out of site sink, implement me!";
-        default:
-            // Unreachable!
-            NPNR_ASSERT(false);
-        }
-    }
 
-    const char * nameOfPip(const SitePip &pip) const {
-        switch(pip.type) {
-        case SitePip::SITE_PIP:
-            return ctx->nameOfPip(pip.pip);
-        case SitePip::SITE_PORT:
-            return ctx->nameOfPip(pip.pip);
-        case SitePip::SOURCE_TO_SITE_PORT:
-            return "source to site port, implement me!";
-        case SitePip::SITE_PORT_TO_SINK:
-            return "site port to sink, implement me!";
-        case SitePip::SITE_PORT_TO_SITE_PORT:
-            return "site port to site port, implement me!";
-        default:
-            // Unreachable!
-            NPNR_ASSERT(false);
-        }
-    }
-
-    const char * nameOfNet(const SiteNetInfo *net) const {
-        return net->net->name.c_str(ctx);
-    }
+    bool debug() const;
 
     bool bindWire(const SiteWire &wire, SiteNetInfo *net) {
         auto result = wire_to_nets.emplace(wire, SiteNetMap{net, 1});
         if(result.first->second.net != net) {
-            if(ctx->debug) {
+            if(debug()) {
                 log_info("Net conflict binding wire %s to net %s, conflicts with net %s\n",
                         nameOfWire(wire), nameOfNet(net), nameOfNet(result.first->second.net));
             }
@@ -628,7 +398,7 @@ struct SiteArch {
         if(!result.second) {
             if(result.first->second.pip != pip) {
                 // Pip conflict!
-                if(ctx->debug) {
+                if(debug()) {
                     log_info("Pip conflict binding pip %s to wire %s, conflicts with pip %s\n",
                             nameOfPip(pip), nameOfWire(dst), nameOfPip(result.first->second.pip));
                 }
@@ -823,28 +593,13 @@ struct SitePipDownhillRange {
     SiteWire site_wire;
     DownhillPipRange pip_range;
 
-    SitePipDownhillRange(const SiteArch *site_arch, SiteWire site_wire) : site_arch(site_arch), site_wire(site_wire) {
-        switch(site_wire.type) {
-        case SiteWire::SITE_WIRE:
-            pip_range = site_arch->ctx->getPipsDownhill(site_wire.wire);
-            break;
-        case SiteWire::OUT_OF_SITE_SOURCE:
-            // No normal pips!
-            break;
-        case SiteWire::OUT_OF_SITE_SINK:
-            // No normal pips!
-            break;
-        case SiteWire::SITE_PORT_SINK:
-            // No normal pips!
-            break;
-        case SiteWire::SITE_PORT_SOURCE:
-            // No normal pips!
-            break;
-        default:
-            // Unreachable!
-            NPNR_ASSERT(false);
+    SitePipDownhillRange(const SiteArch *site_arch, const SiteWire &site_wire) : site_arch(site_arch), site_wire(site_wire) {
+        if(site_wire.type == SiteWire::SITE_WIRE) {
+            init_pip_range();
         }
     }
+
+    void init_pip_range();
 
     SitePipDownhillIterator begin() const {
         SitePipDownhillIterator b;
@@ -1030,28 +785,7 @@ struct SitePipUphillRange {
     SiteWire site_wire;
     UphillPipRange pip_range;
 
-    SitePipUphillRange(const SiteArch *site_arch, SiteWire site_wire) : site_arch(site_arch), site_wire(site_wire) {
-        switch(site_wire.type) {
-        case SiteWire::SITE_WIRE:
-            pip_range = site_arch->ctx->getPipsUphill(site_wire.wire);
-            break;
-        case SiteWire::OUT_OF_SITE_SOURCE:
-            // No normal pips!
-            break;
-        case SiteWire::OUT_OF_SITE_SINK:
-            // No normal pips!
-            break;
-        case SiteWire::SITE_PORT_SINK:
-            // No normal pips!
-            break;
-        case SiteWire::SITE_PORT_SOURCE:
-            // No normal pips!
-            break;
-        default:
-            // Unreachable!
-            NPNR_ASSERT(false);
-        }
-    }
+    SitePipUphillRange(const SiteArch *site_arch, SiteWire site_wire);
 
     SitePipUphillIterator begin() const {
         SitePipUphillIterator b;
@@ -1189,35 +923,7 @@ struct SiteWireIterator {
         return state != other.state || cursor != other.cursor;
     }
 
-    SiteWire operator*() const
-    {
-        WireId wire;
-        PipId pip;
-        SiteWire site_wire;
-        switch(state) {
-        case NORMAL_WIRES:
-            wire.tile = site_arch->site_info->tile;
-            wire.index = cursor;
-            return SiteWire::make(site_arch->site_info, wire);
-        case INPUT_SITE_PORTS:
-            pip = site_arch->input_site_ports.at(cursor);
-            site_wire = SiteWire::make_site_port(site_arch->site_info, pip, site_arch->ctx->getPipSrcWire(pip));
-            NPNR_ASSERT(site_wire.type == SiteWire::SITE_PORT_SOURCE);
-            return site_wire;
-        case OUTPUT_SITE_PORTS:
-            pip = site_arch->output_site_ports.at(cursor);
-            site_wire = SiteWire::make_site_port(site_arch->site_info, pip, site_arch->ctx->getPipDstWire(pip));
-            NPNR_ASSERT(site_wire.type == SiteWire::SITE_PORT_SINK);
-            return site_wire;
-        case OUT_OF_SITE_SOURCES:
-            return site_arch->out_of_site_sources.at(cursor);
-        case OUT_OF_SITE_SINKS:
-            return site_arch->out_of_site_sinks.at(cursor);
-        default:
-            // Unreachable!
-            NPNR_ASSERT(false);
-        }
-    }
+    SiteWire operator*() const;
 };
 
 struct SiteWireRange {
@@ -1228,7 +934,7 @@ struct SiteWireRange {
 
         b.state = SiteWireIterator::BEGIN;
         b.site_arch = site_arch;
-        b.tile_type = &loc_info(site_arch->ctx->chip_info, *site_arch->site_info);
+        b.tile_type = &loc_info(&site_arch->site_info->chip_info(), *site_arch->site_info);
 
         ++b;
         return b;
@@ -1246,7 +952,6 @@ struct SiteWireRange {
 inline SiteWireRange SiteArch::getWires() const {
     return SiteWireRange(this);
 }
-
 
 NEXTPNR_NAMESPACE_END
